% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_documents.R
\name{fetch_wb_documents_json}
\alias{fetch_wb_documents_json}
\title{Fetch a single page from World Bank Documents API}
\usage{
fetch_wb_documents_json(
  base_url = "https://search.worldbank.org/api/v3/wds",
  fl = "id,count,abstracts,authr,docdt,origu,owner,projectid,theme,topic,docty",
  strdate = "2010-01-01",
  enddate = "2026-04-22",
  doc_type = NULL,
  os = 0,
  rows = 200
)
}
\arguments{
\item{base_url}{Character. API endpoint URL.}

\item{fl}{Character. Comma-separated list of fields to return.}

\item{strdate}{Character. Start date in "YYYY-MM-DD" format.}

\item{enddate}{Character. End date in "YYYY-MM-DD" format.}

\item{doc_type}{Character. A vector of document types, to be matched exactly.}

\item{os}{Integer. Offset (starting row).}

\item{rows}{Integer. Number of rows to return.}
}
\value{
A list from the parsed JSON response.
}
\description{
Fetch a single page from World Bank Documents API
}
