% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_portfolioreview.R
\name{render_portfolioreview}
\alias{render_portfolioreview}
\title{A function to render the portfolio review}
\usage{
render_portfolioreview()
}
\description{
A function to render the portfolio review
}
