% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_country_groups}
\alias{wb_country_groups}
\title{World Bank Country Group}
\format{
A data frame with 18 rows and 2 variables:
\describe{
\item{\code{group_name}}{character Group name}
\item{\code{group_category}}{character Group category (e.g., Economic, Region)}
}
}
\source{
https://ddh-openapi.worldbank.org/resources/DR0095333/download
}
\usage{
wb_country_groups
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
