% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_country_list}
\alias{wb_country_list}
\title{World Bank Country and Lending Groups}
\format{
A data frame with 762 rows and 4 variables:
\describe{
\item{\code{country_code}}{character World Bank country code}
\item{\code{country_name}}{character World Bank country name}
\item{\code{group}}{character Country group}
\item{\code{group_code}}{character Country group code}
}
}
\source{
https://ddh-openapi.worldbank.org/resources/DR0095333/download
}
\usage{
wb_country_list
}
\description{
A dataset containing the World Bank's standard country codes, country names, and their respective groups and group codes. This dataset is used to identify countries and their classifications in various World Bank reports and analyses.
}
\keyword{datasets}
